+++ pkg/grpc/service/data_export.go
@@ -16,6 +16,7 @@ import (
 	"github.com/influenzanet/study-service/pkg/exporter"
 	"github.com/influenzanet/study-service/pkg/studyengine"
 	"github.com/influenzanet/study-service/pkg/types"
+	"github.com/influenzanet/study-service/pkg/utils"
 	"google.golang.org/grpc/codes"
 	"google.golang.org/grpc/status"
 
@@ -360,6 +361,42 @@ func (s *studyServiceServer) HasAccessToDownload(t *api_types.TokenInfos, studyK
 	return err
 }
 
+func (s *studyServiceServer) GetResponsesFlatJSONWithPagination(req *api.ResponseExportQuery, stream api.StudyServiceApi_GetResponsesFlatJSONWithPaginationServer) error {
+	buf, err := s.getResponseExportBuffer(req, FLAT_JSON)
+
+	ctx := context.Background()
+	itemCount := s.studyDBservice.GetSurveyResponsesCount(ctx, req.Token.InstanceId, req.StudyKey, req.SurveyKey, req.From, req.Until)
+	pageSize, page, pageCount := utils.ComputePaginationParameter(req.PageSize, req.Page, itemCount)
+	// Send pagination infos
+	infos := &api.PaginatedFile{
+		Data: &api.PaginatedFile_Info{
+			Info: &api.PaginationInfo{
+				ItemCount: itemCount,
+				PageCount: pageCount,
+				PageSize:  pageSize,
+				Page:      page,
+			},
+		},
+	}
+
+	err = stream.Send(infos)
+	if err != nil {
+		return err
+	}
+
+	data := &api.PaginatedFile{
+		Data: &api.PaginatedFile_Chunk{
+			Chunk: buf.Bytes(),
+		},
+	}
+	err = stream.Send(data)
+	if err != nil {
+		return err
+	}
+
+	return nil
+}
+
 // TODO: Test GetResponsesFlatJSON
 func (s *studyServiceServer) GetResponsesFlatJSON(req *api.ResponseExportQuery, stream api.StudyServiceApi_GetResponsesFlatJSONServer) error {
 	buf, err := s.getResponseExportBuffer(req, FLAT_JSON)
@@ -465,7 +502,7 @@ func (s *studyServiceServer) getResponseExportBuffer(req *api.ResponseExportQuer
 		ctx,
 		req.Token.InstanceId, req.StudyKey, req.SurveyKey,
 		req.From, req.Until, func(instanceID, studyKey string, response types.SurveyResponse, args ...interface{}) error {
-			if len(args) != 1 {
+			if len(args) != 3 {
 				return errors.New("[getResponseExportBuffer]: wrong DB method argument")
 			}
 			rExp, ok := args[0].(*exporter.ResponseExporter)
@@ -474,7 +511,7 @@ func (s *studyServiceServer) getResponseExportBuffer(req *api.ResponseExportQuer
 			}
 			return rExp.AddResponse(&response)
 		},
-		responseExporter,
+		responseExporter, req.Page, req.PageSize,
 	)
 	if err != nil {
 		logger.Info.Print(err)