+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Changelog
 
+## ??? - 2023-08-31
+
+### Added
+
+- New db collection for history of study rules, new methods to get, delete and add study rules. 
+- New uploadedAt index model for study rules 
+- New gRPC endpoints: 
+  - `GetStudyRulesHistory`: get study rules history for specified study key that fullfill query criteria with pagination and sorting options 
+  - `GetCurrentStudyRules`: get current study rules for specified study key 
+  - `RemoveStudyRulesVersion`: deletes study rules version with specified id.
+
+### Changed
+
+- Changed gRPC endpoints:
+  - `DeleteStudy`: study rules history is also deleted 
+  - `SaveStudyRules`: new study rules are added in study rules collection 
+
+
+## ??? - 2023-08-01
+
+### Added
+
+- New gRPC endpoint `GetResponsesFlatJSONWithPagination`: stream responses in JSON format with pagination infos.
+
 ## [v1.6.2] - 2023-07-27
 
 ### Added